Move all form step tests to `prescriptionFormConstants.getFormSteps`
